requirements.txt
slack-bolt>=1.21.2
gspread>=6
google-auth>=2
pytz>=2024.1
python-dotenv
